# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SPI API",
    "description": "API for indoor positioning system - final project",
    "version": "1.0.0"
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/buildings/{buildingId}/samples": {
      "get": {
        "x-swagger-router-controller": "samples",
        "operationId": "list",
        "summary": "Get samples of a building by building ID",
        "description": "",
        "tags": ["/buildings/{buildingId}/samples"],
        "parameters": [
          {
            "in": "path",
            "name": "buildingId",
            "description": "The building's id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sample"
              }
            }
          }
        }
      },
      "post": {
        "x-swagger-router-controller": "samples",
        "operationId": "create",
        "summary": "Create a sample for a building",
        "description": "",
        "tags": ["/buildings/{buildingId}/samples"],
        "parameters": [
          {
            "in": "body",
            "name": "payload",
            "description": "Sample payload",
            "schema": {
              "$ref": "#/definitions/Sample"
            }
          },
          {
            "in": "path",
            "name": "buildingId",
            "description": "The building's id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Sample"
            }
          }
        }
      },
      "delete": {
        "x-swagger-router-controller": "samples",
        "operationId": "deleteAll",
        "summary": "Delete all samples for a building",
        "description": "",
        "tags": ["/buildings/{buildingId}/samples"],
        "parameters": [
          {
            "in": "path",
            "name": "buildingId",
            "description": "The building's id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Sample"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Sample": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        },
        "buildingId": {
          "type": "string"
        },
        "floorId": {
          "type": "string"
        },
        "fingerprint": {
          "$ref": "#/definitions/Fingerprint"
        }
      },
      "required": [
        "latitude",
        "longitude",
        "buildingId",
        "floorId",
        "fingerprint"
      ]
    },
    "Fingerprint": {
      "type": "object",
      "additionalProperties": {
        "type": "number"
      },
      "example": {
        "9823472984BFA5C": -35,
        "BD4353453C3453A": -5
      }
    }
  },
  "securityDefinitions": {
    "ApiKey": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  }
}